pyproject.toml
[tool.poetry]
name = "tutorial"
version = "0.1.0"
description = ""
authors = ["Ali Farrokhtala <farrokhtala@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
# psycopg2 = "^2.9.3"
psycopg2-binary = "^2.9.7"
django-filter = "^23.2"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"
ncal = "^0.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pre-commit = "^3.4.0"
colorlog = "^6.7.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
